+++ src/Library/Recipe.cs
@@ -4,35 +4,22 @@
 // </copyright>
 //-------------------------------------------------------------------------
 
-using System;
 using System.Collections;
 
 namespace Full_GRASP_And_SOLID.Library
 {
     public class Recipe
     {
-        private ArrayList steps = new ArrayList();
-
-        public Product FinalProduct { get; set; }
+        public static ArrayList steps = new ArrayList();
 
         public void AddStep(Step step)
         {
-            this.steps.Add(step);
+            steps.Add(step);
         }
 
         public void RemoveStep(Step step)
         {
-            this.steps.Remove(step);
-        }
-
-        public void PrintRecipe()
-        {
-            Console.WriteLine($"Receta de {this.FinalProduct.Description}:");
-            foreach (Step step in this.steps)
-            {
-                Console.WriteLine($"{step.Quantity} de '{step.Input.Description}' " +
-                    $"usando '{step.Equipment.Description}' durante {step.Time}");
-            }
+            steps.Remove(step);
         }
     }
 }
